+++ addons/chatify_selection/lib/selection/view/header.dart
@@ -1,6 +1,5 @@
-import '../../../../../domain/models/messages/message.dart';
-import '../../../bloc/bloc.dart';
-import '../../../../common/animated_flip_counter.dart';
+import 'package:chatify/chatify.dart';
+import 'package:chatify_selection/selection/bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get/get.dart';
@@ -12,15 +11,13 @@ class SelectedMessagesHeader extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final color = Theme.of(context).colorScheme.onPrimary;
-    return BlocSelector<MessagesBloc, MessagesState, Map<String, Message>>(
+    return BlocSelector<SelectionBloc, SelectionState, Map<String, Message>>(
       selector: (state) => state.selectedMessages,
       builder: (context, selecetdMessages) {
         return AnimatedSwitcher(
           duration: const Duration(milliseconds: 300),
           child: selecetdMessages.isEmpty
-              ? const SizedBox.shrink(
-                  key: ValueKey('no-selected-messages'),
-                )
+              ? const SizedBox.shrink(key: ValueKey('no-selected-messages'))
               : Container(
                   decoration: BoxDecoration(
                     color: Theme.of(context).primaryColor,
@@ -32,9 +29,9 @@ class SelectedMessagesHeader extends StatelessWidget {
                       InkWell(
                         highlightColor: Colors.transparent,
                         onTap: () {
-                          context
-                              .read<MessagesBloc>()
-                              .add(MessageDeselectAll());
+                          context.read<SelectionBloc>().add(
+                            SelectionDeselectAll(),
+                          );
                         },
                         child: Container(
                           padding: const EdgeInsetsDirectional.only(
@@ -43,19 +40,13 @@ class SelectedMessagesHeader extends StatelessWidget {
                             top: 12,
                             bottom: 12,
                           ),
-                          child: Icon(
-                            Icons.close,
-                            color: color,
-                          ),
+                          child: Icon(Icons.close, color: color),
                         ),
                       ),
                       AnimatedFlipCounter(
                         value: selecetdMessages.length,
                         duration: const Duration(milliseconds: 200),
-                        textStyle: TextStyle(
-                          fontSize: 16,
-                          color: color,
-                        ),
+                        textStyle: TextStyle(fontSize: 16, color: color),
                       ),
                       Text(
                         ' ${'Selected'.tr}',
@@ -64,18 +55,11 @@ class SelectedMessagesHeader extends StatelessWidget {
                       const Spacer(),
                       IconButton(
                         onPressed: () async {
-                          context
-                              .read<MessagesBloc>()
-                              .add(MessagesDeleteSelected());
+                          context.read<SelectionBloc>().add(SelectionDelete());
                         },
-                        icon: Icon(
-                          Iconsax.trash,
-                          color: color,
-                        ),
+                        icon: Icon(Iconsax.trash, color: color),
                       ),
-                      const SizedBox(
-                        width: 10,
-                      ),
+                      const SizedBox(width: 10),
                     ],
                   ),
                 ),