+++ addons/chatify_selection/lib/selection/view/listener.dart
@@ -1,5 +1,5 @@
-import '../../../../../domain/models/messages/message.dart';
-import '../../../bloc/bloc.dart';
+import 'package:chatify/chatify.dart';
+import 'package:chatify_selection/selection/bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -24,20 +24,22 @@ class _MessagesSelectionListenerState extends State<MessagesSelectionListener> {
   Map<String, Message> initialSelecetdMessages = {};
 
   _detectTapedItem(PointerEvent event) {
-    if (!context.read<MessagesBloc>().state.isSelectionMode) return;
-    final bloc = context.read<MessagesBloc>();
-    final RenderBox box =
-        _key.currentContext!.findAncestorRenderObjectOfType<RenderBox>()!;
+    if (!context.read<SelectionBloc>().state.isSelectionMode) return;
+    final bloc = context.read<SelectionBloc>();
+    final RenderBox box = _key.currentContext!
+        .findAncestorRenderObjectOfType<RenderBox>()!;
     final result = BoxHitTestResult();
     Offset local = box.globalToLocal(event.position);
     if (box.hitTest(result, position: local)) {
       for (final hit in result.path) {
         final target = hit.target;
         if (target is SelectedMessage) {
-          final initialList =
-              Map<String, Message>.from(bloc.state.selectedMessages);
-          final selectedMessages =
-              Map<String, Message>.from(bloc.state.selectedMessages);
+          final initialList = Map<String, Message>.from(
+            bloc.state.selectedMessages,
+          );
+          final selectedMessages = Map<String, Message>.from(
+            bloc.state.selectedMessages,
+          );
           bool isScrollUp = offset.dy > local.dy;
           addedMessages.putIfAbsent(target.index, () => target.message);
           addedMessages.removeWhere(
@@ -52,13 +54,14 @@ class _MessagesSelectionListenerState extends State<MessagesSelectionListener> {
           } else {
             selectedMessages.clear();
             selectedMessages.addAll(
-              addedMessages
-                  .map((key, value) => MapEntry(value.content.id, value)),
+              addedMessages.map(
+                (key, value) => MapEntry(value.content.id, value),
+              ),
             );
             selectedMessages.addAll(initialSelecetdMessages);
           }
           if (initialList.length != selectedMessages.length) {
-            bloc.add(MessageSelectionChanged(selectedMessages));
+            bloc.add(SelectionChanged(selectedMessages));
           }
         }
       }
@@ -67,18 +70,21 @@ class _MessagesSelectionListenerState extends State<MessagesSelectionListener> {
 
   _detectStart(PointerEvent event) {
     addedMessages.clear();
-    initialSelecetdMessages =
-        context.read<MessagesBloc>().state.selectedMessages;
-    final RenderBox box =
-        _key.currentContext!.findAncestorRenderObjectOfType<RenderBox>()!;
+    initialSelecetdMessages = context
+        .read<SelectionBloc>()
+        .state
+        .selectedMessages;
+    final RenderBox box = _key.currentContext!
+        .findAncestorRenderObjectOfType<RenderBox>()!;
     offset = box.globalToLocal(event.position);
     final result = BoxHitTestResult();
     if (box.hitTest(result, position: offset)) {
       for (final hit in result.path) {
         final target = hit.target;
         if (target is SelectedMessage) {
-          isRemove =
-              initialSelecetdMessages.containsKey(target.message.content.id);
+          isRemove = initialSelecetdMessages.containsKey(
+            target.message.content.id,
+          );
         }
       }
     }
@@ -91,7 +97,7 @@ class _MessagesSelectionListenerState extends State<MessagesSelectionListener> {
       onPointerMove: _detectTapedItem,
       onPointerDown: _detectStart,
       onPointerUp: (event) {
-        context.read<MessagesBloc>().add(MessagesSelectionModeChanged(false));
+        context.read<SelectionBloc>().add(SelectionModeChanged(false));
       },
       child: widget.child,
     );